[TM-3606] Move a bunch of utility methods and types out of airtable-entity.ts
