+++ apps/unified-database-service/src/airtable/entities/disturbance.airtable-entity.ts
@@ -1,12 +1,8 @@
 /* istanbul ignore file */
-import {
-  AirtableEntity,
-  associatedValueColumn,
-  ColumnMapping,
-  commonEntityColumns,
-  PolymorphicUuidAssociation
-} from "./airtable-entity";
+import { AirtableEntity } from "./airtable-entity";
 import { Disturbance, DisturbanceReport, Site, SiteReport } from "@terramatch-microservices/database/entities";
+import { associatedValueColumn, commonEntityColumns } from "../util/columns";
+import { ColumnMapping, PolymorphicUuidAssociation } from "../util/types";
 
 const LARAVEL_TYPE_MAPPING: Record<string, PolymorphicUuidAssociation<DisturbanceAssociations>> = {
   [Site.LARAVEL_TYPE]: {
