+++ apps/unified-database-service/src/airtable/entities/funding-programme.airtable-entity.ts
@@ -1,5 +1,7 @@
-import { AirtableEntity, ColumnMapping, commonEntityColumns } from "./airtable-entity";
+import { AirtableEntity } from "./airtable-entity";
 import { FundingProgramme } from "@terramatch-microservices/database/entities";
+import { commonEntityColumns } from "../util/columns";
+import { ColumnMapping } from "../util/types";
 
 const COLUMNS: ColumnMapping<FundingProgramme>[] = [
   ...commonEntityColumns<FundingProgramme>("fundingProgramme"),
