+++ apps/unified-database-service/src/airtable/entities/investment.airtable-entity.ts
@@ -1,6 +1,8 @@
 /* istanbul ignore file */
-import { AirtableEntity, ColumnMapping, commonEntityColumns } from "./airtable-entity";
+import { AirtableEntity } from "./airtable-entity";
 import { Investment, Project } from "@terramatch-microservices/database/entities";
+import { commonEntityColumns } from "../util/columns";
+import { ColumnMapping } from "../util/types";
 
 const COLUMNS: ColumnMapping<Investment>[] = [
   ...commonEntityColumns<Investment>(),
