+++ apps/unified-database-service/src/airtable/entities/nursery.airtable-entity.ts
@@ -1,14 +1,8 @@
-import {
-  AirtableEntity,
-  associatedValueColumn,
-  ColumnMapping,
-  commonEntityColumns,
-  treeAmountRollup,
-  treeDescriptionRollup,
-  UpdateAssociation
-} from "./airtable-entity";
+import { AirtableEntity } from "./airtable-entity";
 import { Nursery, Project, TreeSpecies } from "@terramatch-microservices/database/entities";
 import { groupBy, uniq } from "lodash";
+import { associatedValueColumn, commonEntityColumns, treeAmountRollup, treeDescriptionRollup } from "../util/columns";
+import { ColumnMapping, UpdateAssociation } from "../util/types";
 
 type NurseryAssociations = {
   projectUuid?: string;
