+++ apps/unified-database-service/src/airtable/entities/organisation.airtable-entity.ts
@@ -1,6 +1,8 @@
-import { AirtableEntity, associatedValueColumn, ColumnMapping, commonEntityColumns } from "./airtable-entity";
+import { AirtableEntity } from "./airtable-entity";
 import { Organisation } from "@terramatch-microservices/database/entities";
 import { filter, flatten, uniq } from "lodash";
+import { associatedValueColumn, commonEntityColumns } from "../util/columns";
+import { ColumnMapping } from "../util/types";
 
 type OrganisationAssociations = {
   hqCountryName: string | null;
