+++ apps/unified-database-service/src/airtable/entities/site-polygon.airtable-entity.ts
@@ -1,8 +1,10 @@
 /* istanbul ignore file */
-import { AirtableEntity, associatedValueColumn, ColumnMapping, commonEntityColumns } from "./airtable-entity";
+import { AirtableEntity } from "./airtable-entity";
 import { Disturbance, SitePolygon } from "@terramatch-microservices/database/entities";
 import { isDate, isString, uniq } from "lodash";
 import { isNotNull } from "@terramatch-microservices/database/types/array";
+import { associatedValueColumn, commonEntityColumns } from "../util/columns";
+import { ColumnMapping } from "../util/types";
 
 type SitePolygonAssociations = {
   disturbanceUuid?: string;
