+++ apps/unified-database-service/src/airtable/entities/tracking-entry.airtable-entity.ts
@@ -1,6 +1,8 @@
 import { Tracking, TrackingEntry } from "@terramatch-microservices/database/entities";
-import { AirtableEntity, associatedValueColumn, ColumnMapping } from "./airtable-entity";
+import { AirtableEntity } from "./airtable-entity";
 import { uniq } from "lodash";
+import { associatedValueColumn } from "../util/columns";
+import { ColumnMapping } from "../util/types";
 
 type TrackingEntryAssociations = {
   trackingUuid?: string;
